requirements.txt
cryptography==3.4.8
pycryptodome==3.11.0
